Update rokuencode.php
Modified filesize64() to be Linux-specific (and simpler...not 100% reliable function yet:()
Standardized vars to be the same between script and MythTV vars.
Updated transcode params to take chan and starttime instead of -i and -o.  Script failed previously.
Cosmetic improvements.

TODO: Myth Transcode does seem to degrade the quality marginally.  Still ok, but not supreme.  Not sure why that is.